docs(asset-details-panel):  research possibility to use Props  (#533)
